backend/requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.34.0
pydantic==2.12.5
requests==2.32.5

faiss-cpu==1.13.2
sentence-transformers==5.2.0
numpy>=1.26